hackathon-ai/requirements.txt
python-dotenv
openai==0.28.0
amazon-textract-caller==0.2.1
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.4.5
boto3==1.33.2
botocore==1.33.2
requests==2.31.0
Flask==3.0.0
pdf2image